x86: Update g++.target/i386/pr100885.C
Since long is 32 bits for x32, update g++.target/i386/pr100885.C to cast
__m64 to long long for x32.

	PR target/100885
	* g++.target/i386/pr100885.C (_mm_set_epi64): Cast __m64 to long
	long.

+++ gcc/testsuite/g++.target/i386/pr100885.C
@@ -33,7 +33,7 @@ protected:
   }
 };
 __m128i _mm_set_epi64(__m64 __q0) {
-  __m128i __trans_tmp_5{(long)__q0};
+  __m128i __trans_tmp_5{(long long)__q0};
   return __trans_tmp_5;
 }
 long _mm_storel_epi64___P, Draw_dsts;
